feat: expose date value to timeunit slots
By providing the date value we make it easy to let the user decide how to format the value. E.g. we wanted to show the date in a separate line from the weekday.
Making it possible to provide a formatting option for the timeunit slots would also not suffice, since we then still couldn't format date and weekday differently.